Move test/debug code to module
