+++ src/lib.rs
@@ -1,5 +1,8 @@
 extern crate chrono;
+extern crate clap;
+extern crate reqwest;
 extern crate serde_derive;
+
 #[macro_use]
 extern crate log;
