+++ src/pretty.rs
@@ -1,6 +1,39 @@
 use crate::hydra::{Build, Eval};
 use chrono::NaiveDateTime;
 
+#[cfg(test)]
+use crate::hydra::Search;
+#[cfg(test)]
+use std::fs::File;
+#[cfg(test)]
+use std::io::prelude::*;
+
+#[test]
+// This is useful for developping purpose (this is not a test yet).
+fn builds() -> Result<(), std::io::Error> {
+    let mut file = File::open("data/search-build.json")?;
+    let mut contents = String::new();
+    file.read_to_string(&mut contents)?;
+    let s: Search = serde_json::from_str(&contents)?;
+
+    for b in s.builds {
+        build_pretty_print(&b);
+        println!();
+    }
+    Ok(())
+}
+
+#[test]
+// This is useful for developping purpose (this is not a test yet).
+fn test_eval() -> Result<(), std::io::Error> {
+    let mut file = File::open("data/eval-1525352.json")?;
+    let mut contents = String::new();
+    file.read_to_string(&mut contents)?;
+    let e: Eval = serde_json::from_str(&contents)?;
+    evaluation_pretty_print(&e);
+    Ok(())
+}
+
 pub fn evaluation_pretty_print(e: &Eval) {
     for (k, v) in &e.jobsetevalinputs {
         println!("  {}", k);
